renamed *Info components to Info*
